refactor: move components into separate files
